Move functionality into abstract ProxyFileSystem; address other comments
